// src/models/ionicnative-geoposition.model.ts
/**
 * Représente les données envoyées par le GPS du téléphone
 * Learn more here : Coordinates section on: https://ionicframework.com/docs/native/geolocation/
 */

export class IonicNativeGeoposition{
    latitude: number;
    longitude: number;
    altitude: number;
    speed: number;
    heading: number;
    altitudeAccuracy: number;
    accuracy: number;

}